requirements.txt
lark-oapi>=1.4.0
anthropic>=0.40.0
google-genai>=1.0.0
openai>=1.50.0
pyyaml>=6.0
claude-agent-sdk>=0.1.0
